test(gzip): split async tests and add seeds
