+++ test/ICSharpCode.SharpZipLib.Tests/GZip/GZipTests.cs
@@ -388,32 +388,23 @@ namespace ICSharpCode.SharpZipLib.Tests.GZip
 
 		[Test]
 		[Category("GZip")]
-		public void SmallBufferDecompression()
+		public void SmallBufferDecompression([Values(0, 1, 3)] int seed)
 		{
 			var outputBufferSize = 100000;
-			var inputBufferSize = outputBufferSize * 4;
-			var inputBuffer = Utils.GetDummyBytes(inputBufferSize, seed: 0);
-			
 			var outputBuffer = new byte[outputBufferSize];
+			var inputBuffer = Utils.GetDummyBytes(outputBufferSize * 4, seed);
 
 			using var msGzip = new MemoryStream();
-			using (var gzos = new GZipOutputStream(msGzip))
+			using (var gzos = new GZipOutputStream(msGzip){IsStreamOwner = false})
 			{
-				gzos.IsStreamOwner = false;
-				
 				gzos.Write(inputBuffer, 0, inputBuffer.Length);
-				
-				gzos.Flush();
-				gzos.Finish();
 			}
 
 			msGzip.Seek(0, SeekOrigin.Begin);
-				
-
+	
 			using (var gzis = new GZipInputStream(msGzip))
 			using (var msRaw = new MemoryStream())
 			{
-					
 				int readOut;
 				while ((readOut = gzis.Read(outputBuffer, 0, outputBuffer.Length)) > 0)
 				{
@@ -421,59 +412,13 @@ namespace ICSharpCode.SharpZipLib.Tests.GZip
 				}
 
 				var resultBuffer = msRaw.ToArray();
-
 				for (var i = 0; i < resultBuffer.Length; i++)
 				{
 					Assert.AreEqual(inputBuffer[i], resultBuffer[i]);
 				}
-
-
 			}
 		}
-		
-#if NETCOREAPP3_1_OR_GREATER
-		[Test]
-		[Category("GZip")]
-		[Category("Async")]
-		public async Task SmallBufferDecompressionAsync()
-		{
-			var outputBufferSize = 100000;
-			var inputBufferSize = outputBufferSize * 4;
-			var inputBuffer = Utils.GetDummyBytes(inputBufferSize, seed: 0);
-			
-			var outputBuffer = new byte[outputBufferSize];
 
-			await using var msGzip = new MemoryStream();
-			await using (var gzos = new GZipOutputStream(msGzip){IsStreamOwner = false})
-			{
-				await gzos.WriteAsync(inputBuffer, 0, inputBuffer.Length);
-			}
-
-			msGzip.Seek(0, SeekOrigin.Begin);
-
-
-			using (var gzis = new GZipInputStream(msGzip))
-			await using (var msRaw = new MemoryStream())
-			{
-					
-				int readOut;
-				while ((readOut = await gzis.ReadAsync(outputBuffer, 0, outputBuffer.Length)) > 0)
-				{
-					await msRaw.WriteAsync(outputBuffer, 0, readOut);
-				}
-
-				var resultBuffer = msRaw.ToArray();
-
-				for (var i = 0; i < resultBuffer.Length; i++)
-				{
-					Assert.AreEqual(inputBuffer[i], resultBuffer[i]);
-				}
-
-
-			}
-		}
-#endif
-		
 		/// <summary>
 		/// Should gracefully handle reading from a stream that becomes unreadable after
 		///  all of the data has been read.
